* renamed trainer and advanced trainer files
