+++ example_simple_regression.py
@@ -6,9 +6,8 @@ found that one of the models works well on your data, move to Example2
 import pandas as pd
 
 import healthcareai.common.file_io_utilities as io_utilities
-from healthcareai.trainer import SupervisedModelTrainer
+from healthcareai.supvervised_model_trainer import SupervisedModelTrainer
 import healthcareai.common.write_predictions_to_database as hcaidb
-
 
 
 def main():
