+++ healthcareai/__init__.py
@@ -1,5 +1,5 @@
-from .advanced_trainer import AdvancedSupervisedModelTrainer
-from .trainer import SupervisedModelTrainer
+from .advanced_supvervised_model_trainer import AdvancedSupervisedModelTrainer
+from .supvervised_model_trainer import SupervisedModelTrainer
 
 
 __all__ = [
