+++ healthcareai/tests/test_trained_supervised_model.py
@@ -2,7 +2,7 @@ import unittest
 import pandas as pd
 
 import healthcareai.tests.helpers as helpers
-from healthcareai.trainer import SupervisedModelTrainer
+from healthcareai.supvervised_model_trainer import SupervisedModelTrainer
 
 
 class TestTrainedSupervisedModel(unittest.TestCase):
